A-Packages: * created duke package * organised classes into tasks, parser, commands, exception, ui, and storage packages
