--- src/main/java/Command.java
@@ -1,6 +1,0 @@
-abstract class Command {
-
-    abstract public boolean isExit();
-
-    abstract public void execute(TaskList<Task> tasks, Ui ui, Storage storage);
-}
