+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.storage.Storage;
+import duke.tasks.Task;
+import duke.tasks.TaskList;
+import duke.ui.Ui;
+
 public class DeleteCommand extends Command {
     private final int task;
