+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.storage.Storage;
+import duke.tasks.Task;
+import duke.tasks.TaskList;
+import duke.ui.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public boolean isExit() {
