+++ src/main/java/duke/exceptions/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidCommandException extends DukeException{
     public InvalidCommandException() {
         super();
