+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,17 @@
+package duke.parser;
+
+import duke.commands.Command;
+import duke.commands.DeleteCommand;
+import duke.commands.ExitCommand;
+import duke.commands.ListCommand;
+import duke.commands.MarkCommand;
+import duke.commands.UnmarkCommand;
+import duke.commands.DeadlineCommand;
+import duke.commands.EventCommand;
+import duke.commands.TodoCommand;
+import duke.exceptions.DukeException;
+import duke.exceptions.InvalidCommandException;
+
 import java.util.NoSuchElementException;
 import java.util.StringTokenizer;
