+++ src/main/java/duke/tasks/Event.java
@@ -1,3 +1,7 @@
+package duke.tasks;
+
+import duke.tasks.Task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
