+++ src/main/java/duke/tasks/TaskList.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 import java.util.ArrayList;
 
 public class TaskList<T extends Task> {
