+++ src/main/java/duke/ui/Ui.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 import java.util.Scanner;
 
 public class Ui {
